Move genome folder to shared/state
