+++ src/ensembl/src/content/app/browser/browserSelectors.ts
@@ -2,7 +2,7 @@ import { RootState } from 'src/store';
 import { ChrLocation } from './browserState';
 
 import { getQueryParamsMap } from 'src/global/globalHelper';
-import { getGenomeInfo } from 'src/genome/genomeSelectors';
+import { getGenomeInfo } from 'src/shared/state/genome/genomeSelectors';
 import { getEnsObjectById } from 'src/ens-object/ensObjectSelectors';
 
 export const getBrowserActivated = (state: RootState) =>
