+++ src/ensembl/src/root/rootReducer.ts
@@ -3,7 +3,7 @@ import { connectRouter } from 'connected-react-router';
 
 import browser from '../content/app/browser/browserReducer';
 import drawer from '../content/app/browser/drawer/drawerReducer';
-import genome from '../genome/genomeReducer';
+import genome from '../shared/state/genome/genomeReducer';
 import customDownload from '../content/app/custom-download/state/customDownloadReducer';
 import global from '../global/globalReducer';
 import header from '../header/headerReducer';
