+++ src/ensembl/src/shared/state/genome/genomeActions.ts
@@ -15,7 +15,7 @@ import {
   getGenomeInfoById,
   getGenomeTrackCategories,
   getGenomeKaryotype
-} from 'src/genome/genomeSelectors';
+} from 'src/shared/state/genome/genomeSelectors';
 import { ensureSpeciesIsCommitted } from 'src/content/app/species-selector/state/speciesSelectorActions';
 import { getBrowserActiveGenomeId } from 'src/content/app/browser/browserSelectors';
